do not print a status at the end of each line
The status at the end of each line looks great, unless the step
produced output itself then it looks really bad. This patch changes
such that it does not put a status at the end each line, but does
print out a notice if there is not a step or the step failed at the
bottom of the output.
dont print a status at the end of each line

The status at the end of each line looks great, unless the step
produced output itself then it looks really bad. This patch changes
such that it does not put a status at the end each line, but does
print out a notice if there is not a step or the step failed at the
bottom of the output.
TMP status